requirements.txt
matplotlib
seaborn
pywin32
python-docx
easyocr
Pillow
PyMuPDF
--find-links https://download.pytorch.org/whl/cu126
torch
torchvision
torchaudio